stream_memory: split out bstr logic
This logic can then be reused by other streams that work on static data.
